x86/KVM: PCI: Enable KVM_FEATURE_PCI_GO_MMCONFIG CPUID bit
This feature allows MMCONFIG to be used even to access the base PCI
config space [1]. This means increased performance - one access to MMCONFIG
instead of two conventional accesses to I/O ports.

Q35 makes no distinction in base or extended PCI config access to
MMCONFIG, MMCONFIG is always on, and in case it is is not initialized,
probing of PCI devices will fall back to normal process and use type1
access.

Enable the feature unconditionally.

[1]:

Signed-off-by: Julia Suvorova <jusual@redhat.com>